# src/vmapi/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Virtual Machine Api",
        "title": "Virtual Machine Api",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/vms": {
            "post": {
                "description": "Create a vm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "VMS"
                ],
                "summary": "Create a vm",
                "parameters": [
                    {
                        "description": "vm",
                        "name": "vm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.VMDetail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Res"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Res"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Err"
                        }
                    }
                }
            }
        },
        "/vms/{phy}": {
            "get": {
                "description": "Get all vms",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "VMS"
                ],
                "summary": "Get all vms",
                "parameters": [
                    {
                        "type": "string",
                        "description": "物理机IP",
                        "name": "phy",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.VMS"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Err"
                        }
                    }
                }
            }
        },
        "/vms/{phy}/{vm}": {
            "get": {
                "description": "Get a vm detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "VMS"
                ],
                "summary": "Get a vm detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "物理机IP",
                        "name": "phy",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "虚拟机NAME",
                        "name": "vm",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.VMC"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Err"
                        }
                    }
                }
            },
            "put": {
                "description": "Controller a vm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "VMS"
                ],
                "summary": "Controller a vm",
                "parameters": [
                    {
                        "description": "操作",
                        "name": "operator",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.OP"
                        }
                    },
                    {
                        "type": "string",
                        "description": "物理机IP",
                        "name": "phy",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "虚拟机NAME",
                        "name": "vm",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Res"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Err"
                        }
                    }
                }
            },
            "delete": {
                "description": "Shutdown a vm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "VMS"
                ],
                "summary": "Shutdown a vm",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "强制关机",
                        "name": "Force",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "物理机IP",
                        "name": "phy",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "虚拟机NAME",
                        "name": "vm",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Res"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/model.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Err"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Err": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "integer"
                },
                "message": {
                    "type": "object"
                }
            }
        },
...
        "model.OP": {
            "type": "object",
            "properties": {
                "operator": {
                    "type": "string"
                }
            }
        },
        "model.Res": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "integer"
                },
                "message": {
                    "type": "object"
                }
            }
        },
        "model.VM": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "model.VMC": {
            "type": "object",
            "properties": {
                "cpu": {
                    "type": "integer"
                },
                "dns": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "gateway": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "memory": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "netmask": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "model.VMDetail": {
            "type": "object",
            "properties": {
                "cpu": {
                    "type": "string"
                },
                "gateway": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "mask": {
                    "type": "string"
                },
                "mem": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pyhIp": {
                    "type": "string"
                }
            }
        },
        "model.VMS": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/model.VM"
            }
        }
    }
}